Dungeon Run
Started making the actual dungeon run feature, I added the while loop in order to repeat the code in the middle which would have just run the game once for the first floor. Also the cooldown variable is for the dodge feature where if you have enough agility then the monster can get stunned meaning you get an additional attack on it.