Item Colours
Changes the way that the item colours updated as the dictionary with the colours was in "Common" and in the items.json the rarity was in numbers. Additionally the dictionary messed with the colorama module making Style.BRIGHT + Fore.WHITE into x1/h2/34 or something random like that. (Can't recall off the top of my head"